pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["Jack Surfer <pstnkv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.71.0"
uvicorn = "^0.16.0"
Office365-REST-Python-Client = "^2.3.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
